% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_utils.R
\name{backup_file_path}
\alias{backup_file_path}
\title{Find a backup path}
\usage{
backup_file_path(file)
}
\arguments{
\item{file}{Existing file path.}
}
\value{
A character scalar backup path.
}
\description{
Return the first available \code{.bak}, \code{.bak1}, \code{.bak2}, ... path for a file.
}
