% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colours.R
\name{colour_grid_dims}
\alias{colour_grid_dims}
\title{Compute colour swatch grid dimensions}
\usage{
colour_grid_dims(n, nrow = NULL, ncol = NULL)
}
\arguments{
\item{n}{Number of colours.}

\item{nrow, ncol}{Optional grid dimensions.}
}
\value{
A list with integer elements \code{nrow} and \code{ncol}.
}
\description{
Compute a grid that has enough cells for a given number of colours.
}
