% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colours.R
\name{colour_text_contrast}
\alias{colour_text_contrast}
\title{Choose text colour for contrast}
\usage{
colour_text_contrast(colours)
}
\arguments{
\item{colours}{Character vector of R colours.}
}
\value{
Character vector containing \code{"black"} or \code{"white"}.
}
\description{
Choose black or white text for each background colour using a simple
luminance calculation.
}
