% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{dev_name}
\alias{dev_name}
\title{Current graphics device name}
\usage{
dev_name()
}
\value{
A character scalar. The null device is returned as \code{"null device"}.
}
\description{
Return the name of the currently active graphics device as reported by
\code{\link[grDevices:dev]{grDevices::dev.cur()}}.
}
