% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{named_par_mar}
\alias{named_par_mar}
\title{Named plot margins}
\usage{
named_par_mar()
}
\value{
A named numeric vector with elements \code{bottom}, \code{left}, \code{top}, and
\code{right}.
}
\description{
Return \code{graphics::par("mar")} with stable names.
}
\examples{
plot(1:3, 1:3)
named_par_mar()

}
