% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{named_par_usr}
\alias{named_par_usr}
\title{Named user coordinate limits}
\usage{
named_par_usr()
}
\value{
A named numeric vector with elements \code{left}, \code{right}, \code{bottom},
and \code{top}.
}
\description{
Return \code{graphics::par("usr")} with stable names.
}
\examples{
plot(1:3, 1:3)
named_par_usr()

}
