% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_text.R
\name{nice_text_default_macros_file}
\alias{nice_text_default_macros_file}
\title{Default TeX macro file for \code{nice_text()}}
\usage{
nice_text_default_macros_file()
}
\value{
A character scalar giving the installed file path.
}
\description{
Return the path to the package-supplied default TeX macro file.
}
\examples{
nice_text_default_macros_file()

}
