% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_text.R
\name{nice_text_ignore_commands}
\alias{nice_text_ignore_commands}
\title{Active TeX ignore commands for \code{nice_text()}}
\usage{
nice_text_ignore_commands(ignore.file = NULL, append.ignore = TRUE)
}
\arguments{
\item{ignore.file}{Optional path to a user ignore-command file. If \code{NULL},
only the option \code{earnmisc.tex_ignore_file} is checked.}

\item{append.ignore}{Logical scalar. If \code{TRUE}, include package defaults. If
\code{FALSE}, use only user files.}
}
\value{
A character vector of TeX command names, including the leading
backslash.
}
\description{
Return the active TeX commands removed or simplified by \code{nice_text()} for
non-tikz graphics devices.
}
\details{
Package defaults are read from \code{inst/tex/default-ignore-commands.txt}. If the
option \code{earnmisc.tex_ignore_file} is set, that file is read after the
defaults. If \code{ignore.file} is supplied, it is read after the option file. Set
\code{append.ignore = FALSE} to omit the package defaults.
}
\examples{
nice_text_ignore_commands()

ignore.file <- tempfile()
writeLines("\\\\foo", ignore.file)
nice_text_ignore_commands(ignore.file = ignore.file)

}
