% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okabe_ito.R
\name{oi_alpha}
\alias{oi_alpha}
\title{Add alpha transparency to colours}
\usage{
oi_alpha(colour, alpha)
}
\arguments{
\item{colour}{Character vector of R colour values, such as \code{oi.orange},
\code{"#E69F00"}, or \code{"orange"}.}

\item{alpha}{Numeric scalar in \verb{[0, 1]} giving the alpha transparency to
apply.}
}
\value{
A character vector of alpha-adjusted colours.
}
\description{
Return alpha-adjusted versions of actual R colour values.
}
\examples{
oi_alpha(oi.orange, 0.5)
oi_alpha(c(oi.orange, oi.sky_blue), 0.25)

}
