% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okabe_ito.R
\docType{data}
\name{oi_colours}
\alias{oi_colours}
\alias{oi.black}
\alias{oi.orange}
\alias{oi.sky_blue}
\alias{oi.bluish_green}
\alias{oi.yellow}
\alias{oi.blue}
\alias{oi.vermillion}
\alias{oi.reddish_purple}
\alias{oi.grey}
\alias{oi.amber}
\title{Okabe-Ito colour constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
oi.black

oi.orange

oi.sky_blue

oi.bluish_green

oi.yellow

oi.blue

oi.vermillion

oi.reddish_purple

oi.grey

oi.amber
}
\value{
Character strings containing hexadecimal RGB colours.
}
\description{
Convenient constants for the Okabe-Ito colours and two Okabe-Ito-style
extensions. The first eight colours are the original Okabe-Ito
colourblind-friendly palette. \code{oi.grey} and \code{oi.amber} are extensions used
for convenience.
}
\details{
These constants are actual R colour strings. Use them directly, for example
\code{col = oi.orange}. Quoted strings such as \code{"oi.orange"} are not special R
colours.
}
\examples{
oi.orange
oi.sky_blue

}
\keyword{datasets}
