% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okabe_ito.R
\name{okabe_ito_palette}
\alias{okabe_ito_palette}
\title{Okabe-Ito palette}
\usage{
okabe_ito_palette(n = length(okabe_ito_colours()), alpha = NULL)
}
\arguments{
\item{n}{Integer scalar giving the number of colours to return.}

\item{alpha}{Optional numeric scalar in \verb{[0, 1]} giving the alpha
transparency to apply with \code{\link[grDevices:adjustcolor]{grDevices::adjustcolor()}}. If \code{NULL}, colours
are returned as opaque hexadecimal RGB values.}
}
\value{
A named character vector of \code{n} hexadecimal colour values.
}
\description{
Return the first \code{n} colours from the standard Okabe-Ito palette.
}
\examples{
okabe_ito_palette(3)
okabe_ito_palette(4, alpha = 0.75)

}
