% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plot_metadata}
\alias{plot_metadata}
\title{Base graphics plot metadata}
\usage{
plot_metadata(xlim = NULL, ylim = NULL)
}
\arguments{
\item{xlim}{Optional numeric vector describing the x-axis limits to include
in the returned metadata.}

\item{ylim}{Optional numeric vector describing the y-axis limits to include
in the returned metadata.}
}
\value{
A list with elements \code{xlim}, \code{ylim}, \code{par.usr}, \code{par.mar}, and
\code{par.list}.
}
\description{
Return a small list of metadata about the current base graphics plot.
}
\examples{
plot(1:3, 1:3)
plot_metadata()

}
