% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_pp_plot.R
\name{pp_lines}
\alias{pp_lines}
\title{Overlay phase-plane trajectories}
\usage{
pp_lines(x, ...)
}
\arguments{
\item{x}{Object to overlay.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns updated method-specific plot metadata.
}
\description{
\code{pp_lines()} is an S3 generic for adding phase-plane trajectories to an
existing phase-plane plot.
}
\examples{
time <- seq(0, 2 * pi, length.out = 100)
x <- stats::ts(cbind(sin = sin(time), cos = cos(time)))
info <- pp_plot(x, h.var = "sin", v.var = "cos")
y <- stats::ts(cbind(sin = 0.8 * sin(time), cos = 0.8 * cos(time)))
pp_lines(y, plot.info = info, lty = 2)

}
