% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_pp_plot.R
\name{pp_plot}
\alias{pp_plot}
\title{Plot phase-plane trajectories}
\usage{
pp_plot(x, ...)
}
\arguments{
\item{x}{Object to plot.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns method-specific plot metadata.
}
\description{
\code{pp_plot()} is an S3 generic for phase-plane trajectory plots. Methods should
plot one variable against another and return plot metadata that \code{\link[=pp_lines]{pp_lines()}}
can use for overlays.
}
\examples{
time <- seq(0, 2 * pi, length.out = 100)
x <- stats::ts(cbind(sin = sin(time), cos = cos(time)))
info <- pp_plot(x, h.var = "sin", v.var = "cos")
y <- stats::ts(cbind(sin = 0.8 * sin(time), cos = 0.8 * cos(time)))
pp_lines(y, plot.info = info, lty = 2)

}
