% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colours.R
\name{validate_grid_dimension}
\alias{validate_grid_dimension}
\title{Validate a grid dimension}
\usage{
validate_grid_dimension(x, name)
}
\arguments{
\item{x}{Value to validate.}

\item{name}{Argument name for error messages.}
}
\value{
An integer scalar.
}
\description{
Validate a positive integer scalar grid dimension.
}
