% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colours.R
\name{validate_r_colours}
\alias{validate_r_colours}
\title{Validate R colours}
\usage{
validate_r_colours(colours)
}
\arguments{
\item{colours}{Character vector of colour specifications.}
}
\value{
Invisibly returns \code{colours}.
}
\description{
Validate a character vector by asking base R to convert it to RGB.
}
