% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_utils.R
\name{validate_width_cutoff}
\alias{validate_width_cutoff}
\title{Validate an input-form width cutoff}
\usage{
validate_width_cutoff(width.cutoff)
}
\arguments{
\item{width.cutoff}{Value to validate.}
}
\value{
An integer scalar between 20 and 500.
}
\description{
Validate a value for \code{\link[=deparse]{deparse()}}'s \code{width.cutoff} argument.
}
