% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mts.R
\name{as_mts}
\alias{as_mts}
\title{Convert objects to multivariate time series}
\usage{
as_mts(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object suitable for \code{\link[=mts_plot]{mts_plot()}} and related helpers.
}
\description{
\code{as_mts()} is an S3 generic for converting supported objects to
multivariate time-series objects suitable for \code{\link[=mts_plot]{mts_plot()}} and related
\verb{mts_*} helpers.
}
\details{
Methods should return an object that \code{\link[=mts_plot]{mts_plot()}} can safely treat as
multivariate time-series data, typically an object inheriting from \code{mts}.
}
\examples{
x <- stats::ts(cbind(a = 1:5, b = 6:10))
mts_plot(x)

}
