% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnmisc-package.R
\docType{package}
\name{earnmisc-package}
\alias{earnmisc}
\alias{earnmisc-package}
\title{earnmisc: Miscellaneous Utility Functions}
\description{
\code{earnmisc} provides a small set of general-purpose helper functions used
across related R packages.
}
\details{
The main utility families are:
\itemize{
\item Okabe-Ito colour palette helpers.
\item Small base graphics helpers, including vectorised \code{xys_line()}.
\item Base graphics plot metadata helpers.
\item Lightweight TeX label helpers, including \code{nice_text()}.
\item Tikz graphics device helpers.
}
}
\author{
\strong{Maintainer}: David Earn \email{earn@math.mcmaster.ca}

}
\keyword{internal}
