% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okabe_ito.R
\name{okabe_ito_colours}
\alias{okabe_ito_colours}
\title{Okabe-Ito colours}
\usage{
okabe_ito_colours(alpha = NULL)
}
\arguments{
\item{alpha}{Optional numeric scalar in \verb{[0, 1]} giving the alpha
transparency to apply with \code{\link[grDevices:adjustcolor]{grDevices::adjustcolor()}}. If \code{NULL}, colours
are returned as opaque hexadecimal RGB values.}
}
\value{
A named character vector of hexadecimal colour values.
}
\description{
Return the standard Okabe-Ito colourblind-friendly palette.
}
\examples{
okabe_ito_colours()
okabe_ito_colours(alpha = 0.5)

}
